src/requirements.txt
kedro==0.18.0
kedro-viz==4.5.0
scikit-learn==1.0.2
joblib~=1.0.1 # Required by pandas_profiling
pandas_profiling==3.1.0
scikit-plot==0.3.7
datapane==0.14.0
